+++ temba/channels/handlers.py
@@ -10,6 +10,7 @@ import requests
 import six
 import magic
 import xml.etree.ElementTree as ET
+import logging
 
 from datetime import datetime
 from django.conf import settings
@@ -41,6 +42,8 @@ from temba.utils.text import decode_base64
 from temba.utils.twitter import generate_twitter_signature
 from twilio import twiml
 from .tasks import fb_channel_subscribe, refresh_jiochat_access_tokens
+
+logger = logging.getLogger(__name__)
 
 
 class BaseChannelHandler(View):
@@ -89,6 +92,18 @@ def get_channel_handlers():
         return cls.__subclasses__() + [g for s in cls.__subclasses__() for g in all_subclasses(s)]
 
     return all_subclasses(BaseChannelHandler)
+
+
+class DMarkHandler(BaseChannelHandler):
+    courier_url = r'^dk/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive|status)$'
+    courier_name = 'courier.dk'
+
+    def get(self, request, *args, **kwargs):  # pragma: no cover
+        return HttpResponse("Illegal Method", status_code=405)
+
+    def post(self):  # pragma: no cover
+        logger.error('DMark handling only implemented in courier')
+        return HttpResponse("DMark handling only implemented in Courier.", status_code=401)
 
 
 class TwimlAPIHandler(BaseChannelHandler):
@@ -846,6 +861,7 @@ class InfobipHandler(BaseChannelHandler):
                     sms.status_delivered()
                 elif status in ['REJECTED', 'UNDELIVERABLE']:
                     sms.status_fail()
+                log(sms, 'Status Updated', "SMS Status Updated")
 
             return make_response("SMS Status Updated", status_code=200)
 
@@ -1413,7 +1429,7 @@ class VumiHandler(BaseChannelHandler):
 
         # determine if it's a USSD session message or a regular SMS
         is_ussd = "ussd" in body.get('transport_name', '') or body.get('transport_type', '') == 'ussd'
-        channel_type = Channel.TYPE_VUMI_USSD if is_ussd else Channel.TYPE_VUMI
+        channel_type = 'VMU' if is_ussd else 'VM'
 
         # look up the channel
         channel = Channel.objects.filter(uuid=request_uuid, is_active=True, channel_type=channel_type).exclude(
@@ -1858,6 +1874,9 @@ class MageHandler(BaseChannelHandler):
 
     def post(self, request, *args, **kwargs):
         from temba.triggers.tasks import fire_follow_triggers
+
+        if not settings.MAGE_AUTH_TOKEN:  # pragma: no cover
+            return JsonResponse(dict(error="Authentication not configured"), status=401)
 
         authorization = request.META.get('HTTP_AUTHORIZATION', '').split(' ')