+++ temba/flows/models.py
@@ -194,7 +194,7 @@ class Flow(TembaModel):
     START_MSG_FLOW_BATCH = 'start_msg_flow_batch'
 
     VERSIONS = [
-        "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "10.1"
+        "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "10.1", "10.2"
     ]
 
     name = models.CharField(max_length=64,
@@ -1744,8 +1744,8 @@ class Flow(TembaModel):
         # these fields are the initial state for our flow run
         run_fields = None
         if extra:
-            # we keep 1024 values in @extra for new flow runs because we might be passing the state
-            (normalized_fields, count) = FlowRun.normalize_fields(extra, 1024)
+            # we keep more values in @extra for new flow runs because we might be passing the state
+            (normalized_fields, count) = FlowRun.normalize_fields(extra, settings.FLOWRUN_FIELDS_SIZE * 4)
             run_fields = json.dumps(normalized_fields)
 
         # create all our flow runs for this set of contacts at once
@@ -1773,10 +1773,10 @@ class Flow(TembaModel):
         # if we have some broadcasts to optimize for
         message_map = dict()
         if broadcasts:
-            # create our message context
-            message_context_base = self.build_expressions_context(None, start_msg)
+            # create our expressions context
+            expressions_context_base = self.build_expressions_context(None, start_msg)
             if extra:
-                message_context_base['extra'] = extra
+                expressions_context_base['extra'] = extra
 
             # and add each contact and message to each broadcast
             for broadcast in broadcasts:
@@ -1785,7 +1785,7 @@ class Flow(TembaModel):
 
                 # create the sms messages
                 created_on = timezone.now()
-                broadcast.send(message_context=message_context_base, trigger_send=False,
+                broadcast.send(expressions_context=expressions_context_base, trigger_send=False,
                                response_to=start_msg, status=INITIALIZING, msg_type=FLOW,
                                created_on=created_on, partial_recipients=partial_recipients, run_map=run_map)
 
@@ -2703,10 +2703,13 @@ class FlowRun(models.Model):
         return FlowRun.INVALID_EXTRA_KEY_CHARS.sub('_', key)[:255]
 
     @classmethod
-    def normalize_fields(cls, fields, max_values=256, count=-1):
+    def normalize_fields(cls, fields, max_values=None, count=-1):
         """
         Turns an arbitrary dictionary into a dictionary containing only string keys and values
         """
+        if max_values is None:
+            max_values = settings.FLOWRUN_FIELDS_SIZE
+
         if isinstance(fields, six.string_types):
             return fields[:Value.MAX_VALUE_LEN], count + 1
 
@@ -3002,9 +3005,9 @@ class FlowRun(models.Model):
         contact_runs = cls.objects.filter(is_active=True, contact__in=contacts)
         cls.bulk_exit(contact_runs, exit_type)
 
-    def update_fields(self, field_map, max_values=256):
+    def update_fields(self, field_map):
         # validate our field
-        (field_map, count) = FlowRun.normalize_fields(field_map, max_values)
+        (field_map, count) = FlowRun.normalize_fields(field_map)
 
         if not self.fields:
             self.fields = json.dumps(field_map)
@@ -3575,10 +3578,10 @@ class RuleSet(models.Model):
             for url in urls:
                 from temba.api.models import WebHookEvent
 
-                (value, errors) = Msg.substitute_variables(url, context, org=run.flow.org, url_encode=True)
+                (value, errors) = Msg.evaluate_template(url, context, org=run.flow.org, url_encode=True)
 
                 result = WebHookEvent.trigger_flow_event(run, value, self, msg, action, resthook=resthook,
-                                                         header=header)
+                                                         headers=header)
 
                 # we haven't recorded any status yet, do so
                 if not status_code:
@@ -3620,7 +3623,7 @@ class RuleSet(models.Model):
             # if we have a custom operand, figure that out
             text = None
             if self.operand:
-                (text, errors) = Msg.substitute_variables(self.operand, context, org=run.flow.org)
+                (text, errors) = Msg.evaluate_template(self.operand, context, org=run.flow.org)
             elif msg:
                 text = msg.text
 
@@ -4995,8 +4998,8 @@ class EmailAction(Action):
         from .tasks import send_email_action_task
 
         # build our message from our flow variables
-        (message, errors) = Msg.substitute_variables(self.message, context, org=run.flow.org)
-        (subject, errors) = Msg.substitute_variables(self.subject, context, org=run.flow.org)
+        (message, errors) = Msg.evaluate_template(self.message, context, org=run.flow.org)
+        (subject, errors) = Msg.evaluate_template(self.subject, context, org=run.flow.org)
 
         # make sure the subject is single line; replace '\t\n\r\f\v' to ' '
         subject = regex.sub('\s+', ' ', subject, regex.V0)
@@ -5006,7 +5009,7 @@ class EmailAction(Action):
         for email in self.emails:
             if email.startswith('@'):
                 # a valid email will contain @ so this is very likely to generate evaluation errors
-                (address, errors) = Msg.substitute_variables(email, context, org=run.flow.org)
+                (address, errors) = Msg.evaluate_template(email, context, org=run.flow.org)
             else:
                 address = email
 
@@ -5058,7 +5061,7 @@ class WebhookAction(Action):
     def execute(self, run, context, actionset_uuid, msg, offline_on=None):
         from temba.api.models import WebHookEvent
 
-        (value, errors) = Msg.substitute_variables(self.webhook, context, org=run.flow.org, url_encode=True)
+        (value, errors) = Msg.evaluate_template(self.webhook, context, org=run.flow.org, url_encode=True)
 
         if errors:
             ActionLog.warn(run, _("URL appears to contain errors: %s") % ", ".join(errors))
@@ -5068,7 +5071,7 @@ class WebhookAction(Action):
             for item in self.webhook_headers:
                 headers[item.get('name')] = item.get('value')
 
-        WebHookEvent.trigger_flow_event(run, value, actionset_uuid, msg, self.action, header=headers)
+        WebHookEvent.trigger_flow_event(run, value, actionset_uuid, msg, self.action, headers=headers)
         return []
 
 
@@ -5140,7 +5143,7 @@ class AddToGroupAction(Action):
         if contact:
             for group in self.groups:
                 if not isinstance(group, ContactGroup):
-                    (value, errors) = Msg.substitute_variables(group, context, org=run.flow.org)
+                    (value, errors) = Msg.evaluate_template(group, context, org=run.flow.org)
                     group = None
 
                     if not errors:
@@ -5270,7 +5273,7 @@ class AddLabelAction(Action):
         for label in self.labels:
             if not isinstance(label, Label):
                 contact = run.contact
-                (value, errors) = Msg.substitute_variables(label, context, org=run.flow.org)
+                (value, errors) = Msg.evaluate_template(label, context, org=run.flow.org)
 
                 if not errors:
                     label = Label.label_objects.filter(org=contact.org, name__iexact=value.strip()).first()
@@ -5325,7 +5328,7 @@ class SayAction(Action):
 
         # localize the text for our message, need this either way for logging
         message = run.flow.get_localized_text(self.msg, run.contact)
-        (message, errors) = Msg.substitute_variables(message, context)
+        (message, errors) = Msg.evaluate_template(message, context)
 
         msg = run.create_outgoing_ivr(message, media_url, run.connection)
 
@@ -5362,7 +5365,7 @@ class PlayAction(Action):
         return dict(type=self.TYPE, uuid=self.uuid, url=self.url)
 
     def execute(self, run, context, actionset_uuid, event, offline_on=None):
-        (media, errors) = Msg.substitute_variables(self.url, context)
+        (media, errors) = Msg.evaluate_template(self.url, context)
         msg = run.create_outgoing_ivr(_('Played contact recording'), media, run.connection)
 
         if msg:
@@ -5440,14 +5443,14 @@ class ReplyAction(Action):
                 created_on = None
 
             if msg and msg.id:
-                replies = msg.reply(text, user, trigger_send=False, message_context=context,
+                replies = msg.reply(text, user, trigger_send=False, expressions_context=context,
                                     connection=run.connection, msg_type=self.MSG_TYPE, attachments=attachments,
                                     send_all=self.send_all, created_on=created_on)
             else:
                 # if our run has been responded to or any of our parent runs have
                 # been responded to consider us interactive with high priority
                 high_priority = run.get_session_responded()
-                replies = run.contact.send(text, user, trigger_send=False, message_context=context,
+                replies = run.contact.send(text, user, trigger_send=False, expressions_context=context,
                                            connection=run.connection, msg_type=self.MSG_TYPE, attachments=attachments,
                                            created_on=created_on, all_urns=self.send_all, high_priority=high_priority)
         return replies
@@ -5601,7 +5604,7 @@ class VariableContactAction(Action):
         return variables
 
     def build_groups_and_contacts(self, run, msg):
-        message_context = run.flow.build_expressions_context(run.contact, msg)
+        expressions_context = run.flow.build_expressions_context(run.contact, msg)
         contacts = list(self.contacts)
         groups = list(self.groups)
 
@@ -5619,7 +5622,7 @@ class VariableContactAction(Action):
 
             # other type of variable, perform our substitution
             else:
-                (variable, errors) = Msg.substitute_variables(variable, message_context, org=run.flow.org)
+                (variable, errors) = Msg.evaluate_template(variable, expressions_context, org=run.flow.org)
 
                 # Check for possible contact uuid and use its contact
                 contact_variable_by_uuid = Contact.objects.filter(uuid=variable, org=run.flow.org).first()
@@ -5879,7 +5882,7 @@ class SaveToContactAction(Action):
         # evaluate our value
         contact = run.contact
         user = get_flow_user(run.org)
-        (value, errors) = Msg.substitute_variables(self.value, context, org=run.flow.org)
+        (value, errors) = Msg.evaluate_template(self.value, context, org=run.flow.org)
 
         if contact.is_test and errors:  # pragma: needs cover
             ActionLog.warn(run, _("Expression contained errors: %s") % ', '.join(errors))
@@ -6048,7 +6051,7 @@ class SendAction(VariableContactAction):
 
                 broadcast = Broadcast.create(flow.org, flow.modified_by, self.msg, recipients,
                                              media=self.media, base_language=flow.base_language)
-                broadcast.send(trigger_send=False, message_context=context)
+                broadcast.send(trigger_send=False, expressions_context=context)
                 return list(broadcast.get_messages())
 
             else:
@@ -6061,7 +6064,7 @@ class SendAction(VariableContactAction):
                         unique_contacts.add(contact.pk)
 
                 text = run.flow.get_localized_text(self.msg, run.contact)
-                (message, errors) = Msg.substitute_variables(text, context, org=run.flow.org, partial_vars=True)
+                (message, errors) = Msg.evaluate_template(text, context, org=run.flow.org, partial_vars=True)
 
                 self.logger(run, message, len(unique_contacts))
 
@@ -6514,7 +6517,7 @@ class ContainsTest(Test):
     def evaluate(self, run, sms, context, text):
         # substitute any variables
         test = run.flow.get_localized_text(self.test, run.contact)
-        test, errors = Msg.substitute_variables(test, context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(test, context, org=run.flow.org)
 
         # tokenize our test
         tests = tokenize(test.lower())
@@ -6584,7 +6587,7 @@ class ContainsAnyTest(ContainsTest):
     def evaluate(self, run, sms, context, text):
         # substitute any variables
         test = run.flow.get_localized_text(self.test, run.contact)
-        test, errors = Msg.substitute_variables(test, context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(test, context, org=run.flow.org)
 
         # tokenize our test
         tests = tokenize(test.lower())
@@ -6626,7 +6629,7 @@ class ContainsOnlyPhraseTest(ContainsTest):
     def evaluate(self, run, sms, context, text):
         # substitute any variables
         test = run.flow.get_localized_text(self.test, run.contact)
-        test, errors = Msg.substitute_variables(test, context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(test, context, org=run.flow.org)
 
         # tokenize our test
         tests = tokenize(test.lower())
@@ -6655,7 +6658,7 @@ class ContainsPhraseTest(ContainsTest):
     def evaluate(self, run, sms, context, text):
         # substitute any variables
         test = run.flow.get_localized_text(self.test, run.contact)
-        test, errors = Msg.substitute_variables(test, context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(test, context, org=run.flow.org)
 
         # tokenize our test
         tests = tokenize(test.lower())
@@ -6705,7 +6708,7 @@ class StartsWithTest(Test):
     def evaluate(self, run, sms, context, text):
         # substitute any variables in our test
         test = run.flow.get_localized_text(self.test, run.contact)
-        test, errors = Msg.substitute_variables(test, context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(test, context, org=run.flow.org)
 
         # strip leading and trailing whitespace
         text = text.strip()
@@ -6766,7 +6769,7 @@ class HasDistrictTest(Test):
             return 0, None
 
         # evaluate our district in case it has a replacement variable
-        state, errors = Msg.substitute_variables(self.state, context, org=run.flow.org)
+        state, errors = Msg.evaluate_template(self.state, context, org=run.flow.org)
 
         parent = org.parse_location(state, AdminBoundary.LEVEL_STATE)
         if parent:
@@ -6806,8 +6809,8 @@ class HasWardTest(Test):
         district = None
 
         # evaluate our district in case it has a replacement variable
-        district_name, missing_district = Msg.substitute_variables(self.district, context, org=run.flow.org)
-        state_name, missing_state = Msg.substitute_variables(self.state, context, org=run.flow.org)
+        district_name, missing_district = Msg.evaluate_template(self.district, context, org=run.flow.org)
+        state_name, missing_state = Msg.evaluate_template(self.state, context, org=run.flow.org)
         if (district_name and state_name) and (len(missing_district) == 0 and len(missing_state) == 0):
             state = org.parse_location(state_name, AdminBoundary.LEVEL_STATE)
             if state:
@@ -6858,7 +6861,7 @@ class DateTest(Test):
 
         text = text.replace(' ', "-")
 
-        test, errors = Msg.substitute_variables(self.test, context, org=org)
+        test, errors = Msg.evaluate_template(self.test, context, org=org)
         if not errors:
             date_message = str_to_datetime(text, tz=tz, dayfirst=day_first)
             date_test = str_to_datetime(test, tz=tz, dayfirst=day_first)
@@ -6954,8 +6957,8 @@ class BetweenTest(NumericTest):
         return dict(type=self.TYPE, min=self.min, max=self.max)
 
     def evaluate_numeric_test(self, run, context, decimal_value):
-        min_val, min_errors = Msg.substitute_variables(self.min, context, org=run.flow.org)
-        max_val, max_errors = Msg.substitute_variables(self.max, context, org=run.flow.org)
+        min_val, min_errors = Msg.evaluate_template(self.min, context, org=run.flow.org)
+        max_val, max_errors = Msg.evaluate_template(self.max, context, org=run.flow.org)
 
         if not min_errors and not max_errors:
             try:
@@ -7008,7 +7011,7 @@ class SimpleNumericTest(NumericTest):
 
     # test every word in the message against our test
     def evaluate(self, run, sms, context, text):
-        test, errors = Msg.substitute_variables(str(self.test), context, org=run.flow.org)
+        test, errors = Msg.evaluate_template(str(self.test), context, org=run.flow.org)
 
         text = text.replace(',', '')
         for word in regex.split(r"\s+", text, flags=regex.UNICODE | regex.V0):